+++ packages/shared-core/src/sdk/documents/index.ts
@@ -1,4 +1,4 @@
-export * as applications from "./applications"
+export * as workspaces from "./workspace"
 export * as automations from "./automations"
 export * as users from "./users"
 export * as screens from "./screens"
