+++ packages/shared-core/src/sdk/documents/workspace.ts
